feat(prompt-eval): sub-second wall-clock duration per run, documented
seconds was already measured around the agent call at 1 s resolution; it is
now sub-second and documented as total turn duration (boot + agent assembly +
model and tool calls, judge excluded). Time to first token is not measurable
through the non-streaming CLI call, and transcript timestamps are persist-time
stamps, so neither is inferred.

+++ scripts/prompt-eval/score.py
@@ -321,4 +321,4 @@ if __name__ == "__main__":
     elif cmd == "score":
         doc, case = load_case(sys.argv[2], sys.argv[3])
         run = int(sys.argv[6]) if len(sys.argv) > 6 else 1
-        print(json.dumps(score(doc, case, sys.argv[4], int(sys.argv[5]), run)))
+        print(json.dumps(score(doc, case, sys.argv[4], float(sys.argv[5]), run)))
